Segregation protocol
